Revert "Make DevCommand#getAsJson top level method in BlockScanner.kt"
This reverts commit c077d96a5b2ba21c633abac9f9d20f62c81746f0.
